+++ src/main/java/JesusDeciples/RankingQuiz/api/service/quizDataCenter/state/INIT_NEXT_QUIZ.java
@@ -1,6 +1,6 @@
 package JesusDeciples.RankingQuiz.api.service.quizDataCenter.state;
 
-import JesusDeciples.RankingQuiz.api.webSocket.QuizDataCenter;
+import JesusDeciples.RankingQuiz.api.service.quizDataCenter.QuizDataCenter;
 
 public class INIT_NEXT_QUIZ implements DataCenterState {
     @Override
