+++ src/main/java/JesusDeciples/RankingQuiz/api/service/quizDataCenter/state/ON_QUIZ.java
@@ -1,6 +1,6 @@
 package JesusDeciples.RankingQuiz.api.service.quizDataCenter.state;
 
-import JesusDeciples.RankingQuiz.api.webSocket.QuizDataCenter;
+import JesusDeciples.RankingQuiz.api.service.quizDataCenter.QuizDataCenter;
 
 import java.time.LocalDateTime;
